Initial change for RidePlaceTrack
