+++ src/openrct2/game.h
@@ -25,8 +25,8 @@ enum GAME_COMMAND {
     GAME_COMMAND_SET_RIDE_APPEARANCE,
     GAME_COMMAND_SET_LAND_HEIGHT,
     GAME_COMMAND_TOGGLE_PAUSE,
-    GAME_COMMAND_PLACE_TRACK,
-    GAME_COMMAND_REMOVE_TRACK,
+    GAME_COMMAND_PLACE_TRACK, // GA
+    GAME_COMMAND_REMOVE_TRACK, // GA
     GAME_COMMAND_LOAD_OR_QUIT,
     GAME_COMMAND_CREATE_RIDE,   // GA
     GAME_COMMAND_DEMOLISH_RIDE,
